Secure password and fixed errors
