+++ src/middlewares/isAuthenticaded.middle.js
@@ -4,13 +4,14 @@ export const isAuthenticaded = (req, res, next) => {
   const token = req.header("Authorization").split(" ")[1];
   try {
     const decoded = jwt.verify(token, process.env.SECRET);
-    const { exp: expDate } = decoded;
-    if (Date.now() / 1000 < expDate) {
-      next();
-    } else {
+    next();
+  } catch (err) {
+    if (err.name === "TokenExpiredError") {
       res.status(401).json({ message: "Expired token" });
+    } else if (err.name === "JsonWebTokenError") {
+      res.status(401).json({ message: "Invalid token" });
+    } else {
+      res.status(500).json({ error: true });
     }
-  } catch (error) {
-    res.status(500).json({ error: true });
   }
 };
