+++ src/routes/user.routes.js
@@ -1,10 +1,11 @@
 import express from "express";
 import { register, generateToken } from "../controllers/user.controller.js";
 import { login } from "../middlewares/login.middle.js";
+import { passwordSecurity } from "../middlewares/passwordSecurity.middle.js";
 
 const router = express.Router();
 
 router.post("/login", login, generateToken);
-router.post("/", register);
+router.post("/", passwordSecurity, register);
 
 export default router;
